### 2019-01-17 #1183 @whedon
[PRE REVIEW]: Qiskit VSCode: a tool to enhance the quantum computing programming experience
**Submitting author:** @cbjuan (<a href="http://orcid.org/0000-0003-2045-8329">Juan Cruz-Benito</a>)
**Repository:** <a href="https://github.com/Qiskit/qiskit-vscode/" target ="_blank">https://github.com/Qiskit/qiskit-vscode/</a>
**Version:** v0.4.1
