### 2019-01-17 #1183 @cbjuan
👋 @danielskatz! Thanks for your comments. I fixed the ORCID, and now @whedon can compile the PDF.

About the article proof, maybe we need to fix some titles in the references to meet some formal considerations. For example, the reference:
```
Research, I. (2018b). The ibm quantum experience. https://www.research.ibm.com/
ibm-q/.
```
should be (note the fixes in the author and the capitalization in the title):
```
IBM Research (2018b). The IBM Quantum Experience. https://www.research.ibm.com/
ibm-q/.
```
If possible, I would like to know how to fix these minor issues considering how PDF compiler handles that.

As prospective reviewers (I got them from the list) I would like to point out: @krischer, @mamrehn, and @malmaud. They have been selected mainly regarding their declared knowledge on Typescript and Python.
