### 2019-01-17 #1183 @danielskatz
typically in bibtex files, you add {}s around anything you want to stay in the same case, for example:

  title={The {IBM Q} experience and {QISKit} open-source quantum computing software}
