### 2019-01-17 #1183 @danielskatz
👋 @krischer, @mamrehn, and @malmaud - would a couple of you be willing to review this JOSS submission?
