### 2019-01-17 #1183 @malmaud
Yes, I'm happy to review this. Full disclosure though,  I have worked at IBM research although not in this field and I do not know any of the authors.
