### 2019-01-17 #1183 @danielskatz
@malmaud - Given what you've said, I don't see your previous work at IBM as an issue, though it's good to note.  I'll assign you as a reviewer, but we won't start the review until we get a second reviewer.  @cbjuan - if you have other suggestions outside the JOSS volunteer reviewer community, we're always willing to bring in more people.
