### 2019-01-19 #1183 @cbjuan
Sure, sorry for the delay. From outside JOSS reviewers, I would like to recommend @victorvicpal. And from JOSS reviewers (if needed) I point out to @pragyansmita
