### 2019-01-21 fe2e29f
Removing the support for Qiskit terra v0.5.X
