### 2019-01-22 8f26b46
Ensuring the extension's proper running  when using qiskit 0.6
