### 2019-01-22 4f2a223
Updating executeQASM to use properly Qiskit v0.7
