### 2019-01-22 4bf0673
Updating getBackendStatus, listLocalBackends and listRemoteBackends to use Qiskit v0.7
