+++ client/resources/qiskitScripts/qiskitTools.py
@@ -66,7 +66,7 @@ class QiskitTools(object):
                            hub=None, group=None, project=None):
 
         if (version.parse(__version__) >= version.parse("0.6") and
-                (version.parse(__version__) < version.parse("0.7"))):
+                (version.parse(__version__) <= version.parse("0.7"))):
 
             if (hub is None or group is None or project is None):
                 IBMQ.enable_account(apiToken, url)
@@ -79,26 +79,26 @@ class QiskitTools(object):
 
         else:
             raise QiskitUnsupportedVersion(
-                'Qiskit-terra version must be v0.6')
+                'Qiskit-terra version must be v0.6 or v0.7')
 
         return backs
 
     def listLocalBackends(self):
 
         if (version.parse(__version__) >= version.parse("0.6") and
-                (version.parse(__version__) < version.parse("0.7"))):
+                (version.parse(__version__) <= version.parse("0.7"))):
             backs = [backend.name() for backend in Aer.backends()]
 
         else:
             raise QiskitUnsupportedVersion(
-                'Qiskit-terra version must be v0.6')
+                'Qiskit-terra version must be v0.6 or v0.7')
 
         return backs
 
     def getBackendStatus(self, back, apiToken, url,
                          hub=None, group=None, project=None):
         if (version.parse(__version__) >= version.parse("0.6") and
-                (version.parse(__version__) < version.parse("0.7"))):
+                (version.parse(__version__) <= version.parse("0.7"))):
 
             if (hub is None or group is None or project is None):
                 IBMQ.enable_account(apiToken, url)
@@ -111,7 +111,7 @@ class QiskitTools(object):
 
         else:
             raise QiskitUnsupportedVersion(
-                'Qiskit-terra version must be v0.6')
+                'Qiskit-terra version must be v0.6 or v0.7')
 
     def createDeviceStatus(self, back):
         if (version.parse(__version__) >= version.parse("0.6") and
