### 2019-01-22 e6f020a
Improving qiskit imports
