### 2019-01-22 1dde1f5
Adding comments to reflect how to ignore some or every warning
