### 2019-01-22 22b9749
Rewriting warnings.showwarning to mask all warnings raised within the script
