### 2019-01-22 86ebaf7
Improving the warning filtering and avoiding the auto-imports linting
