+++ client/resources/qiskitScripts/executeQASM.py
@@ -3,9 +3,9 @@
 # This source code is licensed under the Apache License, Version 2.0 found in
 # the LICENSE.txt file in the root directory of this source tree.
 
-import warnings
-from marshmallow.warnings import ChangedInMarshmallow3Warning
-warnings.simplefilter('ignore', category=ChangedInMarshmallow3Warning)
+import warnings  # noqa
+from marshmallow.warnings import ChangedInMarshmallow3Warning  # noqa
+warnings.simplefilter('ignore', category=ChangedInMarshmallow3Warning)  # noqa
 
 from qiskitTools import QiskitTools
 import os
