+++ client/resources/qiskitScripts/listRemoteBackends.py
@@ -3,6 +3,10 @@
 # This source code is licensed under the Apache License, Version 2.0 found in
 # the LICENSE.txt file in the root directory of this source tree.
 
+import warnings  # noqa
+from marshmallow.warnings import ChangedInMarshmallow3Warning  # noqa
+warnings.simplefilter('ignore', category=ChangedInMarshmallow3Warning)  # noqa
+
 from qiskit import __version__
 from packaging import version
 import argparse
