+++ client/resources/qiskitScripts/qiskitTools.py
@@ -1,3 +1,8 @@
+# Copyright (c) 2018, IBM.
+#
+# This source code is licensed under the Apache License, Version 2.0 found in
+# the LICENSE.txt file in the root directory of this source tree.
+
 import json
 from qiskit import __version__
 from IBMQuantumExperience import IBMQuantumExperience
