### 2019-01-23 c6dd3b1
Removing unneeded comments
