### 2019-01-23 89fdd14
Fixing get queue status command for Qiskit >=v0.7
