### 2019-01-23 07217ea
Fixing createDeviceStatus method used to get IBMQ devices status
