### 2019-01-23 d39c389
Adding a Support Decision Record to gather the decision behind the maintaining of different Qiskit versions
