+++ docs/support/SDR-001.md
@@ -1,0 +1,19 @@
+# 1. The extension should support only the current Qiskit-Terra version and a previous one.
+
+Date: 2019-01-23
+
+## Status
+
+Accepted
+
+## Context
+
+Qiskit-vscode is tied to Qiskit libraries, mainly to Qiskit Terra. Qiskit Terra is currently below version 1.0 and experiences many changes in the API each time is upgraded. We must decide how many different versions of the Qiskit Terra's API we support.
+
+## Decision
+
+As decided by the authors and part of the Qiskit core team, we will support the current version of Qiskit Terra and a previous one. Example: currently, the last version is v0.7.X and we will support v0.7.X and v0.6.X.
+
+## Consequences
+
+When Qiskit Terra achieves a stable version, we will need to review this policy.
