### 2019-01-23 8e371df
Adding PR 61 to the CHANGELOG
