### 2019-01-23 968245d
Minor fix in CHANGELOG
