### 2019-01-23 2caeba9
Adding the support for Qiskit Terra v0.7.X to the README
