### 2019-01-25 #1188 @malmaud
I'm currently blocked since the VScode extension appears to automatically install the latest version of qiskit terra (0.7), and yet only works with 0.6.

I'm not sure how to manually downgrade since the qiskit system is split into several Python packages and I don't know what version each of them needs to be to result in a usable configuration (and I don't really have the time to figure it out). Users shouldn't be expected to manually use pip to downgrade to a configuration of qiskit packages that works when they first install this extension and I don't think this can be accepted until the installation is seamless.

A Github issue does indicate an effort is underway to support terra to 0.7 so I would suggest suspending this review until that gets completed.
