### 2019-01-28 #1188 @cbjuan
Hi @malmaud, @danielskatz

We have almost finished the support of the new version of Qiskit (v0.7) https://github.com/Qiskit/qiskit-vscode/pull/61. We hope to release the new version tomorrow. Let's wait for that to make easier the review. Thanks for your support.
