### 2019-01-29 e92d886
Minor fix on CHANGELOG
