### 2019-02-05 e25ec81
Adding Qiskit citation to the README
