### 2019-02-05 a07487a
Updating "authors" header to "authors and citation"
