### 2019-02-05 f538aa5
Updating CHANGELOG to include PR #63
