### 2019-02-05 7073838
Bump version to v0.5.0
