### 2019-02-05 fb2e55f
Minor fixes in CHANGELOG
