### 2019-02-05 #1188 @cbjuan
Sorry for the delay, but we finally released the Qiskit-vscode version v0.5.0 today https://github.com/Qiskit/qiskit-vscode/releases/tag/v0.5.0. You can use it for the review.

https://marketplace.visualstudio.com/items?itemName=qiskit.qiskit-vscode
