### 2019-02-08 82002e7
Updating CHANGELOG for v0.5.1
