+++ README.md
@@ -1,6 +1,7 @@
 # Qiskit VSCode Extension
 
 [![Build Status](https://www.travis-ci.org/Qiskit/qiskit-vscode.svg?branch=master)](https://www.travis-ci.org/Qiskit/qiskit-vscode)
+[![status](http://joss.theoj.org/papers/7687665c24744f6cc3aa30666a70b219/status.svg)](http://joss.theoj.org/papers/7687665c24744f6cc3aa30666a70b219)
 
 > Simplifying Qiskit to make developing quantum circuits and applications faster.
 
