### 2019-02-08 eef86e9
Updating JOSS badge to use HTTPS
