### 2019-02-08 Qiskit/qiskit-vscode#68 @malmaud
Getting error on startup
### Informations

-   **VSCode version**: 1.31.0
-   **Qiskit [Terra / Aqua / other] version**: 0.7.2
-   **Qiskit VSCode Extension version**: 0.5.1
-   **Operating system**: OS X 10.14.2

I'm seeing the attached screenshot when I open up a blank Python buffer.
![screen shot 2019-02-08 at 4 08 55 pm](https://user-images.githubusercontent.com/987837/52506308-2a989700-2bbc-11e9-9001-af9520fd6160.png)
