### 2019-02-08 #1188 @malmaud
Blocked by https://github.com/Qiskit/qiskit-vscode/issues/68
