### 2019-02-11 Qiskit/qiskit-vscode#68 @cbjuan
Hi @malmaud

Can you paste the code you are using to test? I would like to reproduce the issue in my local machine.

Thanks
