### 2019-02-11 Qiskit/qiskit-vscode#68 @malmaud
No code - I just installed the extension, created a new python file, and saw the error siren above that the extension host terminated. I then clicked the 'open developer tools' button on that error popup.
