### 2019-02-11 Qiskit/qiskit-vscode#68 @cbjuan
That's super weird. Sorry about that, we are going to take a look at this.
