### 2019-02-20 Qiskit/qiskit-vscode#68 @lgj2911
Hi, the same thing happens to me, but the error appears after I install the qiskit VSC addon... And I followed all the instructions to install qiskit and the extension...

I am attaching a screen capture
