### 2019-02-26 Qiskit/qiskit-vscode#68 @cbjuan
Hi @malmaud and @lgj2911. We are into this; it seems to be related to this issue https://github.com/Microsoft/vscode/issues/68024

We are trying to figure out [how to address it](https://github.com/Microsoft/vscode/issues/68024#issuecomment-461707138). We will inform you when we solve it.
