### 2019-02-26 #1188 @cbjuan
Hi @danielskatz. We are trying to fix it. It seems to be an error related to the new version of VSCode. We hope to have a solution ASAP (not only for this review but also for the other users reporting the error).
