### 2019-03-05 b7cd8ec
Updating everything to meet new dependencies and their versions
