+++ server/src/qasm/antlr/QasmParserVisitor.ts
@@ -1,7 +1,7 @@
 // Generated from QasmParser.g4 by ANTLR 4.6-SNAPSHOT
 
 
-import { ParseTreeVisitor } from 'antlr4ts/tree/ParseTreeVisitor';
+import { ParseTreeVisitor } from 'antlr4ts';
 
 import { CodeContext } from './QasmParser';
 import { HeadersContext } from './QasmParser';
