+++ server/src/qasm/antlr/QasmParser.ts
@@ -1,28 +1,24 @@
 // Generated from QasmParser.g4 by ANTLR 4.6-SNAPSHOT
 
-
-import { ATN } from 'antlr4ts/atn/ATN';
-import { ATNDeserializer } from 'antlr4ts/atn/ATNDeserializer';
-import { FailedPredicateException } from 'antlr4ts/FailedPredicateException';
-import { NotNull } from 'antlr4ts/Decorators';
-import { NoViableAltException } from 'antlr4ts/NoViableAltException';
-import { Override } from 'antlr4ts/Decorators';
-import { Parser } from 'antlr4ts/Parser';
-import { ParserRuleContext } from 'antlr4ts/ParserRuleContext';
-import { ParserATNSimulator } from 'antlr4ts/atn/ParserATNSimulator';
-import { ParseTreeListener } from 'antlr4ts/tree/ParseTreeListener';
-import { ParseTreeVisitor } from 'antlr4ts/tree/ParseTreeVisitor';
-import { RecognitionException } from 'antlr4ts/RecognitionException';
-import { RuleContext } from 'antlr4ts/RuleContext';
-import { RuleVersion } from 'antlr4ts/RuleVersion';
-import { TerminalNode } from 'antlr4ts/tree/TerminalNode';
-import { Token } from 'antlr4ts/Token';
-import { TokenStream } from 'antlr4ts/TokenStream';
-import { Vocabulary } from 'antlr4ts/Vocabulary';
-import { VocabularyImpl } from 'antlr4ts/VocabularyImpl';
-
-import * as Utils from 'antlr4ts/misc/Utils';
-
+import { 
+	ATN,
+    ATNDeserializer,
+    FailedPredicateException,
+    NoViableAltException,
+    Parser,
+    ParserRuleContext,
+    ParserATNSimulator,
+    ParseTreeListener,
+    ParseTreeVisitor,
+    RecognitionException,
+    RuleContext,
+    RuleVersion,
+    TerminalNode,
+    Token,
+    TokenStream,
+    Vocabulary,
+    VocabularyImpl
+} from 'antlr4ts';
 import { QasmParserListener } from './QasmParserListener';
 import { QasmParserVisitor } from './QasmParserVisitor';
 
@@ -131,26 +127,26 @@ export class QasmParser extends Parser {
 	];
 	public static readonly VOCABULARY: Vocabulary = new VocabularyImpl(QasmParser._LITERAL_NAMES, QasmParser._SYMBOLIC_NAMES, []);
 
-	@Override
-	@NotNull
+	/// @Override
+	/// @NotNull
 	public get vocabulary(): Vocabulary {
 		return QasmParser.VOCABULARY;
 	}
 
-	@Override
+	/// @Override
 	public get grammarFileName(): string { return "QasmParser.g4"; }
 
-	@Override
+	/// @Override
 	public get ruleNames(): string[] { return QasmParser.ruleNames; }
 
-	@Override
+	/// @Override
 	public get serializedATN(): string { return QasmParser._serializedATN; }
 
 	constructor(input: TokenStream) {
 		super(input);
 		this._interp = new ParserATNSimulator(QasmParser._ATN, this);
 	}
-	@RuleVersion(0)
+	@RuleVersion<CodeContext>(0)
 	public code(): CodeContext {
 		let _localctx: CodeContext = new CodeContext(this._ctx, this.state);
 		this.enterRule(_localctx, 0, QasmParser.RULE_code);
@@ -210,7 +206,7 @@ export class QasmParser extends Parser {
 		}
 		return _localctx;
 	}
-	@RuleVersion(0)
+	@RuleVersion<HeadersContext>(0)
[…]
