+++ server/src/qasm/antlr/QasmParser.ts
@@ -1,71 +1,75 @@
-// Generated from QasmParser.g4 by ANTLR 4.6-SNAPSHOT
+// Generated from QasmParser.g4 by ANTLR 4.7.3-SNAPSHOT
 
-import { 
-	ATN,
-    ATNDeserializer,
-    FailedPredicateException,
-    NoViableAltException,
-    Parser,
-    ParserRuleContext,
-    ParserATNSimulator,
-    ParseTreeListener,
-    ParseTreeVisitor,
-    RecognitionException,
-    RuleContext,
-    RuleVersion,
-    TerminalNode,
-    Token,
-    TokenStream,
-    Vocabulary,
-    VocabularyImpl
-} from 'antlr4ts';
-import { QasmParserListener } from './QasmParserListener';
-import { QasmParserVisitor } from './QasmParserVisitor';
+
+import { ATN } from "antlr4ts/atn/ATN";
+import { ATNDeserializer } from "antlr4ts/atn/ATNDeserializer";
+import { FailedPredicateException } from "antlr4ts/FailedPredicateException";
+import { NotNull } from "antlr4ts/Decorators";
+import { NoViableAltException } from "antlr4ts/NoViableAltException";
+import { Override } from "antlr4ts/Decorators";
+import { Parser } from "antlr4ts/Parser";
+import { ParserRuleContext } from "antlr4ts/ParserRuleContext";
+import { ParserATNSimulator } from "antlr4ts/atn/ParserATNSimulator";
+import { ParseTreeListener } from "antlr4ts/tree/ParseTreeListener";
+import { ParseTreeVisitor } from "antlr4ts/tree/ParseTreeVisitor";
+import { RecognitionException } from "antlr4ts/RecognitionException";
+import { RuleContext } from "antlr4ts/RuleContext";
+//import { RuleVersion } from "antlr4ts/RuleVersion";
+import { TerminalNode } from "antlr4ts/tree/TerminalNode";
+import { Token } from "antlr4ts/Token";
+import { TokenStream } from "antlr4ts/TokenStream";
+import { Vocabulary } from "antlr4ts/Vocabulary";
+import { VocabularyImpl } from "antlr4ts/VocabularyImpl";
+
+import * as Utils from "antlr4ts/misc/Utils";
+
+import { QasmParserListener } from "./QasmParserListener";
+import { QasmParserVisitor } from "./QasmParserVisitor";
 
 
 export class QasmParser extends Parser {
-	public static readonly Comment=1;
-	public static readonly WhiteSpace=2;
-	public static readonly Real=3;
-	public static readonly Int=4;
-	public static readonly QasmDescriptor=5;
-	public static readonly Include=6;
-	public static readonly Qelib=7;
-	public static readonly Qreg=8;
-	public static readonly Creg=9;
-	public static readonly Clean=10;
-	public static readonly U=11;
-	public static readonly Cx=12;
-	public static readonly Sin=13;
-	public static readonly Cos=14;
-	public static readonly Tan=15;
-	public static readonly Exp=16;
-	public static readonly Ln=17;
-	public static readonly Sqrt=18;
-	public static readonly Measure=19;
-	public static readonly Barrier=20;
-	public static readonly Reset=21;
-	public static readonly Opaque=22;
-	public static readonly If=23;
-	public static readonly Equals=24;
-	public static readonly Assign=25;
-	public static readonly Semi=26;
-	public static readonly Comma=27;
-	public static readonly LeftCurlyBrace=28;
[…]
