+++ server/src/qiskit/antlr/Python3Lexer.ts
@@ -1,4 +1,4 @@
-// Generated from Python3.g4 by ANTLR 4.6-SNAPSHOT
+// Generated from Python3.g4 by ANTLR 4.7.3-SNAPSHOT
 
 
   import { Token } from 'antlr4ts/Token';
@@ -6,115 +6,122 @@
   import { Python3Parser } from './Python3Parser';
 
 
-import {
-    ATN,
-    ATNDeserializer,
-    CharStream,
-    Lexer,
-    LexerATNSimulator,
-    RuleContext,
-    Vocabulary,
-    VocabularyImpl
-} from 'antlr4ts';
+import { ATN } from "antlr4ts/atn/ATN";
+import { ATNDeserializer } from "antlr4ts/atn/ATNDeserializer";
+import { CharStream } from "antlr4ts/CharStream";
+import { Lexer } from "antlr4ts/Lexer";
+import { LexerATNSimulator } from "antlr4ts/atn/LexerATNSimulator";
+import { NotNull } from "antlr4ts/Decorators";
+import { Override } from "antlr4ts/Decorators";
+import { RuleContext } from "antlr4ts/RuleContext";
+import { Vocabulary } from "antlr4ts/Vocabulary";
+import { VocabularyImpl } from "antlr4ts/VocabularyImpl";
 
-import { Utils } from 'antlr4ts';
+import * as Utils from "antlr4ts/misc/Utils";
 
 
 export class Python3Lexer extends Lexer {
-	public static readonly DEF=1;
-	public static readonly RETURN=2;
-	public static readonly RAISE=3;
-	public static readonly FROM=4;
-	public static readonly IMPORT=5;
-	public static readonly AS=6;
-	public static readonly GLOBAL=7;
-	public static readonly NONLOCAL=8;
-	public static readonly ASSERT=9;
-	public static readonly IF=10;
-	public static readonly ELIF=11;
-	public static readonly ELSE=12;
-	public static readonly WHILE=13;
-	public static readonly FOR=14;
-	public static readonly IN=15;
-	public static readonly TRY=16;
-	public static readonly FINALLY=17;
-	public static readonly WITH=18;
-	public static readonly EXCEPT=19;
-	public static readonly LAMBDA=20;
-	public static readonly OR=21;
-	public static readonly AND=22;
-	public static readonly NOT=23;
-	public static readonly IS=24;
-	public static readonly NONE=25;
-	public static readonly TRUE=26;
-	public static readonly FALSE=27;
-	public static readonly CLASS=28;
-	public static readonly YIELD=29;
-	public static readonly DEL=30;
-	public static readonly PASS=31;
-	public static readonly CONTINUE=32;
-	public static readonly BREAK=33;
-	public static readonly NEWLINE=34;
-	public static readonly NAME=35;
-	public static readonly STRING_LITERAL=36;
-	public static readonly BYTES_LITERAL=37;
[…]
