### 2019-03-05 35fd770
Adding PR #72 to the changelog
