+++ CHANGELOG.md
@@ -22,6 +22,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### ✏️ Changed
 
+-   Updating dependencies to meet qiskit-chemistry and current supported versions ([#72](https://github.com/Qiskit/qiskit-vscode/pull/72) by [@cbjuan](https://github.com/cbjuan))
+
 ### 👾 Security
 
 ### ❌ Removed
