### 2019-03-05 272c539
Adding PR 74 to the CHANGELOG
