### 2019-03-05 583a933
Fixing wanings related to tsConfigFile and enableTSDiagnostics
