### 2019-03-05 8c0317f
Fixing warning related to "ts-jest/dist/preprocessor.js"
