+++ server/jest.config.js
@@ -6,7 +6,7 @@ module.exports = {
     },
     moduleFileExtensions: ['ts', 'js'],
     transform: {
-        '^.+\\.(ts|tsx)$': './node_modules/ts-jest/preprocessor.js'
+        '^.+\\.(ts|tsx)$': 'ts-jest'
     },
     testMatch: ['**/*.test.(ts|js)'],
     roots: ['./test/'],
