### 2019-03-05 b68d811
Adding PR #76 to the CHANGELOG
