+++ CHANGELOG.md
@@ -22,6 +22,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### ✏️ Changed
 
+-   Update the Jest config to avoid warnings in tests ([#76](https://github.com/Qiskit/qiskit-vscode/pull/76) by [@cbjuan](https://github.com/cbjuan))
+
 ### 👾 Security
 
 ### ❌ Removed
