### 2019-03-06 139fcac
Updating the release to v0.5.2 in the CHANGELOG
