+++ CHANGELOG.md
@@ -20,6 +20,16 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### 🐛 Fixed
 
+### ✏️ Changed
+
+### 👾 Security
+
+### ❌ Removed
+
+## [v0.5.2] - 2019-03-06
+
+### 🐛 Fixed
+
 -   Improving checks for Qiskit versions ([#74](https://github.com/Qiskit/qiskit-vscode/pull/74) by [@cbjuan](https://github.com/cbjuan))
 -   Updating dependencies to fix the error on startup #68 ([#70](https://github.com/Qiskit/qiskit-vscode/pull/70) by [@cbjuan](https://github.com/cbjuan))
 
@@ -27,10 +37,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 -   Update the Jest config to avoid warnings in tests ([#76](https://github.com/Qiskit/qiskit-vscode/pull/76) by [@cbjuan](https://github.com/cbjuan))
 -   Updating dependencies to meet qiskit-chemistry and current supported versions ([#72](https://github.com/Qiskit/qiskit-vscode/pull/72) by [@cbjuan](https://github.com/cbjuan))
-
-### 👾 Security
-
-### ❌ Removed
 
 ## [v0.5.1] - 2019-02-08
 
@@ -307,7 +313,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 -   Re-styling charts ([#97](https://github.ibm.com/IBMQuantum/qiskit-vscode/pull/97) by [@cbjuan](https://github.com/cbjuan))
 -   Extension size reduction ([#76](https://github.ibm.com/IBMQuantum/qiskit-vscode/pull/76) by [@ydarias](https://github.com/ydarias))
 
-[unreleased]: https://github.com/Qiskit/qiskit-vscode/compare/v0.5.1...HEAD
+[unreleased]: https://github.com/Qiskit/qiskit-vscode/compare/v0.5.2...HEAD
+[v0.5.2]: https://github.com/Qiskit/qiskit-vscode/compare/v0.5.1...v0.5.2
 [v0.5.1]: https://github.com/Qiskit/qiskit-vscode/compare/v0.5.0...v0.5.1
 [v0.5.0]: https://github.com/Qiskit/qiskit-vscode/compare/v0.4.1...v0.5.0
 [v0.4.1]: https://github.com/Qiskit/qiskit-vscode/compare/v0.4.0...v0.4.1
