### 2019-03-06 Qiskit/qiskit-vscode#68 @cbjuan
Hi @malmaud and @lgj2911

We are releasing the version v0.5.2. In the tests done, the extension works as expected. Please, reopen this issue if you find again the error on startup.

Thanks
