### 2019-03-06 #1188 @cbjuan
We solved the issue that was blocking the review in the new version [0.5.2](https://github.com/Qiskit/qiskit-vscode/releases/tag/v0.5.2).

Hope it allows the review as expected.
