### 2019-03-06 #1188 @arfon
@malmaud, @naberek - seems like @cbjuan has fixed the installation issues. Could you please try again?
