+++ client/src/activationUtils.ts
@@ -137,80 +137,51 @@ export namespace ActivationUtils {
             ),
             vscode.commands.registerCommand('qiskit-vscode.runQiskitCode', () =>
                 CommandExecutor.execPythonActiveEditor().then(codeResult => {
-                    let resultProvider = new ResultProvider();
-                    vscode.workspace.registerTextDocumentContentProvider('qiskit-preview-result', resultProvider);
-                    let previewUri = vscode.Uri.parse(`qiskit-preview-result://authority/result-preview`);
-
                     const codeFile = vscode.window.activeTextEditor.document;
                     codeFile.save();
-                    resultProvider.displayContent(
-                        VizManager.createViz(codeFile.fileName.toString(), codeResult),
-                        previewUri
+                    const panel = vscode.window.createWebviewPanel(
+                        'ExecutionResultQiskit', // Identifies the type of the webview. Used internally
+                        'Execution result - Qiskit', // Title of the panel displayed to the user
+                        vscode.ViewColumn.Two, // Editor column to show the new webview panel in.
+                        {
+                            enableScripts: true,
+                            retainContextWhenHidden: true
+                        } // Enable JS & retain context
                     );
-
-                    vscode.commands
-                        .executeCommand(
-                            'vscode.previewHtml',
-                            previewUri,
-                            vscode.ViewColumn.Two,
-                            'Execution result - Qiskit'
-                        )
-                        .then(
-                            _success => {},
-                            reason => {
-                                QLogger.error(`Error: ${reason}`, this);
-                                vscode.window.showErrorMessage(reason);
-                            }
-                        );
+                    // And set its HTML content
+                    panel.webview.html = VizManager.createViz(codeFile.fileName.toString(), codeResult);
                 })
             ),
             vscode.commands.registerCommand('qiskit-vscode.runQASMCode', () =>
                 CommandExecutor.execQasmActiveEditor(executeQASMScript).then(codeResult => {
-                    let resultProvider = new ResultProvider();
-                    vscode.workspace.registerTextDocumentContentProvider('qasm-preview-result', resultProvider);
-                    let previewUri = vscode.Uri.parse(`qasm-preview-result://authority/result-preview`);
                     let execPath = Util.getOSDependentPath(executeQASMScript);
-                    resultProvider.displayContent(VizManager.createViz(execPath, codeResult), previewUri);
-
-                    vscode.commands
-                        .executeCommand(
-                            'vscode.previewHtml',
-                            previewUri,
-                            vscode.ViewColumn.Two,
-                            'Execution result - QASM'
-                        )
-                        .then(
-                            _success => {},
-                            reason => {
-                                QLogger.error(`Error: ${reason}`, this);
-                                vscode.window.showErrorMessage(reason);
-                            }
-                        );
+                    const panel = vscode.window.createWebviewPanel(
+                        'ExecutionResultQASM', // Identifies the type of the webview. Used internally
+                        'Execution result - QASM', // Title of the panel displayed to the user
+                        vscode.ViewColumn.Two, // Editor column to show the new webview panel in.
+                        {
+                            enableScripts: true,
+                            retainContextWhenHidden: true
+                        } // Enable JS & retain context
+                    );
+                    // And set its HTML content
+                    panel.webview.html = VizManager.createViz(execPath, codeResult);
                 })
             ),
             vscode.commands.registerCommand('qiskit-vscode.discoverLocalBackends', () =>
                 CommandExecutor.execPythonFile(localBackendsScript, []).then(localBackends => {
-                    let resultProvider = new ResultProvider();
-                    vscode.workspace.registerTextDocumentContentProvider('qiskit-localBackends-result', resultProvider);
[…]
