### 2019-03-11 1d5c882
Removing (now) unneeded resultProvider
