### 2019-03-11 34daf69
Updating the changelog for v0.5.3
