### 2019-03-14 #1188 @malmaud
Getting https://github.com/Qiskit/qiskit-vscode/issues/81 at the moment.  I'm guessing some kind of version mismatch in my installed qiskit packages, although I just followed the installation prompts for qiskit that the extension itself offered with VSCode.
