### 2019-03-14 Qiskit/qiskit-vscode#81 @cbjuan
Thanks for detecting it, @malmaud. We are fixing it in the PR #82
