### 2019-03-14 bdd962e
Adding PR 82 to the CHANGELOG
