### 2019-03-14 Qiskit/qiskit-vscode#81 @cbjuan
Released new version [v0.5.4](https://github.com/Qiskit/qiskit-vscode/releases/tag/v0.5.4). Please update your extension in VSCode, it should solve the issue and let you use the extension as expected.
