### 2019-03-14 #1188 @cbjuan
Thanks @malmaud

It should be solved in [v0.5.4](https://github.com/Qiskit/qiskit-vscode/releases/tag/v0.5.4). Please update your extension in VSCode.
