### 2019-03-14 Qiskit/qiskit-code-assistant-vscode#84 @malmaud
'Get queue status' unreliable
Informations
VSCode version: 1.32.2
Qiskit [Terra / Aqua / other] version: 0.5.4
qiskit 0.8.0
terra 0.7.1
aqua 0.4.1
Qiskit VSCode Extension version: 0.5.3
Operating system: OS X 10.14.3

### What is the current behavior?
Running the 'get queue status' command is unreliable. Around 60% of the time, it successfully opens up a new panel and shows the queue. But the other 40% of the time, I see a 'Running' message, then the running message disappears, but no new tab with the results open up. Seems like this might be some kind of race condition.
