### 2019-03-14 #1188 @malmaud
That worked. Running into https://github.com/Qiskit/qiskit-vscode/issues/84 now though.

At this point, I can't really be involved in more rounds of debugging. I'd be happy to either write a review now, given how everything stands, or review at some later point after things have perhaps matured a bit.
