### 2019-03-15 #1188 @arfon
@cbjuan - given it's not been possible for @malmaud to even get to the point of installing your software yet, I think we should pause this review until you have taken some additional steps to verify that the installation process is a smooth one for new users.

For example, if you have access to a second machine, perhaps try following your own installation instructions to see how far you can get? Alternatively, perhaps a colleague could try installing the package too?

JOSS reviewers can only be expected to do so much debugging of installation challenges and I think we've reached this limit.
