### 2019-03-19 #1188 @cbjuan
@malmaud thanks for your input. In this case, we are not aware of this behavior, and we want to test it in depth.
@arfon at the moment @malmaud can install the extension and use it (at least partially). In any case, thanks for the review and we agree it is the moment to pause it. We are sorry for the review process; several changes in the VSCode APIs led us to provide an unstable version for the review with the new versions of the IDE.

We will keep in touch in the future.
